pyproject.toml
[tool.poetry]
name = "Delivery-bot"
version = "0.0.1"
description = ""
authors = ["Maxim Nelin <nelin.max0@gmail.com>"]

[tool.poetry.scripts]

[tool.poetry.dependencies]
python = "3.10.x"
numpy = "^1.22.4"
pandas = "^1.4.2"
sklearn = "^0.0"
matplotlib = "^3.5.2"
aiogram = "^2.20"
python-dotenv = "^0.20.0"
scipy = "^1.8.1"
pylint = "2.11.1"

[tool.poetry.dev-dependencies]
black = { version = "*", allow-prereleases = true }
pylint = "^2.11.1"
commitizen = "^2.20.0"
pre-commit = "^2.15.0"
jupyter = "^1.0.0"
pydot = "^1.4.2"
graphviz = "^0.20"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[pipenv]
allow-prereleases = false

[tool.black]
line-length = 120